Add comment on MakeInteraction and CheckCollision according to EventObject problem

+++ Classes/GameObjectConfig.h
@@ -1,6 +1,6 @@
 #pragma once
 
-enum OBJECT_TYPE : short
+enum OBJECT_TYPE : int
 {
 	NONE,
 	INTERACTIVE,
